change callback setup in HTML UI example + projects configurations

+++ Wolf Engine 2.0/Pipeline.cpp
@@ -185,7 +185,7 @@ Wolf::Pipeline::Pipeline(const ShaderCreateInfo& computeShaderInfo, std::span<Vk
 {
 	VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
 	pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-	pipelineLayoutInfo.setLayoutCount = descriptorSetLayouts.size();
+	pipelineLayoutInfo.setLayoutCount = static_cast<uint32_t>(descriptorSetLayouts.size());
 	pipelineLayoutInfo.pSetLayouts = descriptorSetLayouts.data();
 	pipelineLayoutInfo.pushConstantRangeCount = 0;
